requirements.txt
numpy==1.20.1
python-telegram-bot==13.3
pandas==1.2.2
requests==2.25.1